% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{genInitialConnections}
\alias{genInitialConnections}
\title{Initialize Adjacency Matrix Connections}
\usage{
genInitialConnections(initNC, N, sparseMatrix = TRUE)
}
\arguments{
\item{N}{integer; population size}

\item{sparseMatrix}{boolean; Whether the returned adjacency matrix should be formatted
as a sparseMatrix}
}
\description{
Initialize Adjacency Matrix Connections
}
