% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genRandomBetaMat}
\alias{genRandomBetaMat}
\title{Generate Random Beta matrix for Effective ij Contact Rates}
\usage{
genRandomBetaMat(N)
}
\arguments{
\item{N}{integer; population size}
}
\description{
Generate Random Beta matrix for Effective ij Contact Rates
}
