% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genRandomNetworkConnections}
\alias{genRandomNetworkConnections}
\title{Generate Random Network for Connections}
\usage{
genRandomNetworkConnections(N, rho, initNC)
}
\arguments{
\item{N}{integer; population size}

\item{rho}{numeric; network rewiring probability}

\item{initNC}{integer; neighborhood size as defined by \link{igraph::neighborhood.size}}
}
\description{
Generate Random Network for Connections
}
