% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.GillespieSIRne}
\alias{is.GillespieSIRne}
\title{Check if GillespieSIRne S3 Class}
\usage{
is.GillespieSIRne(x)
}
\arguments{
\item{x}{GillespieSIRne Simulation}
}
\description{
Overload is: function for determining if object is of class GillespieSIRne
}
