% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.GillespieSIRne}
\alias{print.GillespieSIRne}
\title{print GillespieSIRne S3 Class}
\usage{
\method{print}{GillespieSIRne}(x, ...)
}
\arguments{
\item{x}{GillespieSIRne Simulation}

\item{...}{further arguments passed to or from other methods.}
}
\description{
overload print() function to print summary only
}
