% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rewireNEnodes}
\alias{rewireNEnodes}
\title{Update Adjacency Matrix Connections}
\usage{
rewireNEnodes(adjmat = NULL, N)
}
\arguments{
\item{N}{integer; population size}
}
\description{
Update Adjacency Matrix Connections
}
\details{
Will sample a single pair of connections
}
