% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_Gillespie_SIR}
\alias{sim_Gillespie_SIR}
\title{Gillespie Simulation}
\usage{
sim_Gillespie_SIR(
  Iseed = 1,
  N = 1000,
  beta = genRandomBetaMat(N),
  dur_I = 1,
  term_time = 500
)
}
\arguments{
\item{Iseed}{integer; Number of initial infections in the population}

\item{N}{integer; population size}

\item{beta}{numeric vector; Effective contact rate for S_i I_j elements}

\item{dur_I}{numeric; Duration of infection}

\item{term_time}{numeric; time termination condition}
}
\description{
Assuming:
\itemize{
closed population
}
}
