% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.GillespieSIRne}
\alias{summary.GillespieSIRne}
\title{Summary of GillespieSIRne S3 Class}
\usage{
\method{summary}{GillespieSIRne}(object, ...)
}
\arguments{
\item{object}{GillespieSIRne Simulation}

\item{...}{further arguments passed to or from other methods.}
}
\description{
overload summary() function.
}
