% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyout.GillespieSIRne}
\alias{tidyout.GillespieSIRne}
\title{Tidy Out Sim}
\usage{
\method{tidyout}{GillespieSIRne}(x)
}
\arguments{
\item{x}{GillespieSIRne Simulation}
}
\description{
Function for taking output of SIR NE and lifting it over
}
