% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzztradGillespie.R
\name{tradsim_Gillespie_SIR}
\alias{tradsim_Gillespie_SIR}
\title{Traditional Gillespie Simulation}
\usage{
tradsim_Gillespie_SIR(Iseed = 1, N = 10, beta = 1, dur_I = 1, term_time = 500)
}
\arguments{
\item{Iseed}{integer; Number of initial infections in the population}

\item{N}{integer; population size}

\item{beta}{numeric vector; Probability of transmission given contact for S_i I_j elements}

\item{dur_I}{numeric; Duration of infection}

\item{term_time}{numeric; time termination condition}
}
\description{
Internal function for checking mass action of dynamic gillespie
notably drawing uncoupled times to events and events
}
