% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateNetworkConnections}
\alias{updateNetworkConnections}
\title{Update  Network Connections}
\usage{
updateNetworkConnections(adjmat = NULL, rho = 0.5)
}
\arguments{
\item{adjmat}{matrix; adjacency matrix from network of connections}

\item{rho}{numeric; network rewiring probability}
}
\description{
Update  Network Connections
}
