% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genInitialConnections}
\alias{genInitialConnections}
\title{Initialize Adjacency Matrix Connections}
\usage{
genInitialConnections(N, rho)
}
\arguments{
\item{N}{integer; population size}

\item{rho}{numeric; network rewiring rate}
}
\description{
Initialize Adjacency Matrix Connections
}
