% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{genInitialConnections}
\alias{genInitialConnections}
\title{Initialize Adjacency Matrix Connections}
\usage{
genInitialConnections(initNC, N)
}
\arguments{
\item{initNC}{NC integer; initial connections; NB only used if an initial matrix is not provided.}

\item{N}{integer; population size}
}
\description{
Initialize Adjacency Matrix Connections
}
